review: address third-round Copilot + CodeRabbit feedback on PR #260
Code/doc fixes:
- ring_transport/tests.rs: seed `MissionObjective.status` with the
  correct `STATUS_COMPLETED` constant (it had been set to the unrelated
  `MISSION_COMPLETED` constant from the mission-status set; currently
  masked because the gate inspects `MissionInstance.status` only, but
  semantically wrong).
- executor/tests.rs: rename `cross_world_teleport_action_with_unknown_
  entity_is_silent_no_op` to `..._dispatches_gate_travel`; the doc
  comment said "no GateTravel" but the assertion actually requires one,
  matching ring's `Effect::TeleportCrossWorld` arm.
- ring_transport/regions.rs: rewrite the `required_mission_id` doc to
  describe the real gate check (`MissionInstance.status ==
  MISSION_COMPLETED` via `mission_gate_satisfied`); the previous wording
  referenced a non-existent `mission_<id>_status` counter.
- event_dispatch/mod.rs: remove stale "(legacy comment block)" pointer
  next to the `mission` submodule re-exports.

Test guard upgrade:
- event_dispatch/mission.rs: replace the no-op-passing positive tests
  for `fire_mission_completed` / `fire_mission_accepted` with
  observable `IncrementCounter` assertions. The previous bodies would
  have continued passing if the dispatcher were stubbed to
  `Box::pin(async {})`.

File split (clearing the 700-line cap):
- event_dispatch/mod.rs went from 797 lines to 297 by extracting:
    lifecycle.rs   (fire_player_loaded, fire_entity_death)
    interaction.rs (fire_interact_tag, fire_interact_template)
    region.rs      (fire_enter_region, fire_exit_region, fire_teleport_in)
    inventory.rs   (fire_item_use, fire_item_equipped)
    dialog.rs      (fire_dialog_open, fire_dialog_choice)
- mod.rs keeps `fire_chain_by_id` (no trigger family fits) and the
  existing tests; re-exports preserve the `crate::cell::content::fire_*`
  surface so external callers don't change.
